A very painful refactoring process
